src: extract Runtime-owned HeapManager
Move SVM heap reservation and handle aperture management into a
HeapManager owned by Runtime. Reset system heap state alongside other
heap manager fields on tear-down.
